requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
baukit @ git+https://github.com/davidbau/baukit@9d51abd51ebf29769aecc38c4cbef459b731a36e
beartype==0.14.1
better-abc==0.0.3
blis==0.7.11
boto3==1.34.134
botocore==1.34.134
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.5
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cupy-cuda12x==13.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
exceptiongroup==1.2.1
executing==2.0.1
factscore==0.2.0
faiss-gpu==1.7.2
fancy-einsum==0.0.3
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
graphviz==0.20.3
huggingface-hub==0.23.3
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jaxtyping==0.2.31
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lit==18.1.7
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
matplotlib-venn==0.11.10
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.10
optree==0.11.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
preshed==3.0.9
progressbar==2.5
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pysqlite-binary==0.5.1.3380300
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12.1
tenacity==8.4.1
thinc==8.1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==1.13.1
torchopt==0.7.3
torchtyping==0.1.4
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformer-lens==2.1.0
transformers==4.42.1
triton==2.0.0
typeguard==2.13.3
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.3
wasabi==1.1.3
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
